requirements.txt
# Core ML and Image Processing
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Medical Imaging
pydicom>=2.4.0
nibabel>=5.1.0
SimpleITK>=2.2.0

# AI and ML
transformers>=4.30.0
openai>=1.0.0
tensorflow>=2.13.0
keras>=2.13.0

# Web Framework
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
grad-cam>=1.4.0

# Medical Literature
biopython>=1.81.0
requests>=2.31.0

# Database and Storage
sqlite3
pymongo>=4.5.0
redis>=4.6.0

# PDF Generation
reportlab>=4.0.0
PyPDF2>=3.0.0

# Authentication and Security
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0

# Utilities
pandas>=2.0.0
tqdm>=4.65.0
python-multipart>=0.0.6
fastapi>=0.100.0
uvicorn>=0.23.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and Logging
loguru>=0.7.0
prometheus-client>=0.17.0

# Deployment
gunicorn>=21.2.0
docker>=6.1.0
